refactor: Refactor report command into dedicated module
Move generate command implementation from commands.ts to
report.command.ts as part of CLI structure reorganization for better
maintainability
