+++ src/cli/index.ts
@@ -1,8 +1,11 @@
 #!/usr/bin/env node
+
 import 'dotenv/config';
-import { Command } from 'commander';
-import { setupCommands } from './commands.js';
 import pkg from '../../package.json' with { type: 'json' };
+
+import { Command } from 'commander';
+
+import { report } from '@/cli/commands/report.command.js';
 
 const program = new Command();
 
@@ -10,6 +13,6 @@ program
   .version(pkg.version)
   .description(pkg.description);
 
-setupCommands(program);
+report(program);
 
 program.parse(process.argv);
